requirements.txt
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.3
bleach==6.3.0
build==1.4.0
CacheControl==0.14.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cleo==2.1.0
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
crashtest==0.4.1
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
distlib==0.4.0
dulwich==0.24.10
et_xmlfile==2.0.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.3
findpython==0.7.1
fonttools==4.61.1
fqdn==1.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
installer==0.7.0
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
jaraco.classes==3.4.0
jaraco.context==6.0.2
jaraco.functools==4.4.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
keyring==25.7.0
kiwisolver==1.4.9
lark==1.3.1
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mistune==3.2.0
more-itertools==10.8.0
msgpack==1.1.2
narwhals==2.15.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.5.1
notebook_shim==0.2.4
numpy==2.4.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pbs-installer==2025.12.17
pillow==12.1.0
pkginfo==1.12.1.2
platformdirs==4.5.1
plotly==6.5.1
poetry==2.2.1
poetry-core==2.2.1
prometheus_client==0.23.1
prompt_toolkit==3.0.52
psutil==7.2.1
psycopg2-binary==2.9.11
pure_eval==0.2.3
pycparser==2.23
Pygments==2.19.2
pyparsing==3.3.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytz==2025.2
pywin32-ctypes==0.2.3
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
RapidFuzz==3.14.3
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
Send2Trash==2.0.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.1
stack-data==0.6.3
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tomlkit==0.13.3
tornado==6.5.4
traitlets==5.14.3
trove-classifiers==2025.12.1.14
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.3
uv==0.10.10
virtualenv==20.36.1
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
widgetsnbextension==4.0.15
zstandard==0.25.0
passlib[bcrypt]
PyJWT
python-dotenv
fastapi
uvicorn
sqlalchemy
